Update esell.sql
added some comment for category and subcategory tables and changed type_of_categories in table article_types to null since user will enter types, and subtypes, and it will be visible later when user post ad, will be redirected to update ad of types category. If it is not labeled, it will be categoreized to other category, if field is null, when menu is made.